+++ ExtendableEnums.Serialization.Newtonsoft.UnitTests/ExtendableEnumContractResolverTests/ResolveContractConverterShould.cs
@@ -41,4 +41,3 @@ public class ResolveContractConverterShould
         result.Should().BeNull();
     }
 }
-